+++ src/devicecontroller/NoOpDeviceController.ts
@@ -1,6 +1,7 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
+import Destroyable from '../destroyable/Destroyable';
 import DeviceChangeObserver from '../devicechangeobserver/DeviceChangeObserver';
 import DeviceControllerBasedMediaStreamBroker from '../mediastreambroker/DeviceControllerBasedMediaStreamBroker';
 import NoOpMediaStreamBroker from '../mediastreambroker/NoOpMediaStreamBroker';
@@ -67,5 +68,12 @@ export default class NoOpDeviceController
 
   getVideoInputQualitySettings(): VideoQualitySettings | null {
     return null;
+  }
+}
+
+export class DestroyableNoOpDeviceController extends NoOpDeviceController implements Destroyable {
+  destroyed = false;
+  async destroy(): Promise<void> {
+    this.destroyed = true;
   }
 }